+++ packages/landing-scripts/config/webpack.config.base.js
@@ -1,5 +1,4 @@
 const paths = require('./paths');
-const WebpackMessages = require('webpack-messages');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 const isProduction = process.env.NODE_ENV === 'production';
@@ -48,7 +47,14 @@ module.exports = {
           {
             test: /\.js$/,
             exclude: /(node_modules|bower_components)/,
-            use: 'babel-loader',
+            use: {
+              loader: 'babel-loader',
+              options: {
+                cacheDirectory: true,
+                babelrc: false,
+                presets: [require.resolve('babel-preset-stage-1')],
+              },
+            },
           },
           {
             exclude: /\.(js|html|css|scss)$/,
@@ -67,9 +73,6 @@ module.exports = {
     new MiniCssExtractPlugin({
       filename: '[name].css',
       chunkFilename: '[id].css',
-    }),
-    new WebpackMessages({
-      name: 'landing-page',
     }),
   ],
 };