+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.4.3 (June 8, 2018)
+
+### New features
+
+- Serve HTTPS with `--https` option enabled.
+- Generate the SSL certificate per instance.
+
+### Fixes
+
+- Use `react-dev-utils` to format webpack messages, removed deprecated `.plugin` webpack usage.
+
+
+## 0.3.1 (June 6, 2018)
+
+### New features
+
+- Allow access via IP address
+- Use enviroment port and host
+
+### Fixes
+
+- Error on executing babel ([#9](https://github.com/kevindantas/create-landing-page/issues/9))
+
 ## 0.3.0 (May 28, 2018)
 
 ### New features